Added user checks for duplicate user creation, and delete user

+++ src/user/user.service.ts
@@ -20,20 +20,12 @@ export class UserService {
 				id,
 			},
 		});
+		
 		return user;
 	}
 
 	// Create a user
 	async registerUser(data: Prisma.UserCreateInput): Promise<User | Error> {
-		//find out if there is a duplicate user
-		const get = await this.prisma.user.findMany({
-			where: {
-				email: data.email,
-			},
-		});
-
-		console.log(get);
-
 		const {
 			id,
 			email,
@@ -45,11 +37,21 @@ export class UserService {
 			passwordConf,
 		} = data;
 
+		const safeEmail = email.toLowerCase();
+		//find out if there is a duplicate user
+		const get = await this.prisma.user.findUnique({
+			where: {
+				email: safeEmail,
+			},
+		});
+
+		
+
 		if (password !== passwordConf) {
 			throw new Error("Passwords provided are not matching...");
 		}
 
-		const safeEmail = email.toLowerCase();
+		
 		const hashedPassword = await hash(password, 10);
 		const hashedPasswordConf = hashedPassword;
 
@@ -63,10 +65,17 @@ export class UserService {
 			password: hashedPassword,
 			passwordConf: hashedPasswordConf,
 		};
-		const res = await this.prisma.user.create({
-			data: payload,
-		});
-		return res;
+		
+		///Avoids duplicate value(email) if the exist already
+		if (get === null){
+			const res = await this.prisma.user.create({
+				data: payload,
+			});
+			
+			return res;
+		}
+		
+		return new Error("User has an account already.");
 	}
 
 	// Update a user
@@ -100,11 +109,20 @@ export class UserService {
 	}
 
 	// delete a user
-	async deleteUser(id: string): Promise<User> {
+	async deleteUser(id: string): Promise<User | Error> {
+
+		const res = await this.user(id).then((data)=> {
+			return data
+		})
+		
+		if( res === null){
+			return new Error (`The user with ${id}, does not exist`);
+		}
+
 		return this.prisma.user.delete({
-			where: {
+			where:{
 				id,
-			},
+			}
 		});
 	}
 }
